Misc fixes

+++ WRF/run_wrf.sh
@@ -48,7 +48,7 @@ module load WRF/4.4-foss-2022a-dmpar
 echo "Backup real.exe logs"
 mkdir -p LOGS
 cp -af namelist.input LOGS/namelist.input.real
-cp -af rsl.error.0000 LOGS/rsl.error.0000.real
+if [ -f rsl.error.0000 ]; then cp -af rsl.error.0000 LOGS/rsl.error.0000.real; fi
 ls -al > LOGS/ls_real.log
 
 ########################################################
